requirements.txt
Django==3.1.4
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
pytest-django==4.1.0
pytest-cov==2.11.1
djoser==2.1.0
django-cors-headers==3.7.0